feat: make app use Electron api and support desktop notification
